Attempt to improve the results of the synthesizing step in analyze_document_iteratively which produces the dictionary final_entities by adding specific prompt instructions according to the document type:
Here is the discussion sent to gemini:

Ok but final_analysis is passed to addaptive chunking and this uses the structural_units determined in the variable final_analysis. This is surely a problem then if final_analysis doesn't correctly divide the books by chapters - which currently it is not consistently doing. It occurs to me that we could have an additional instruction or set of instructions which could be included dependent on the document type.

I have added a DocumentType document type enum in a new file enums_and_constants.py. which (for now) has these values
class DocumentType(Enum):
    NOVEL = "Novel"
    BIOGRAPHY = "Biography"
    JOURNAL_ARTICLE = "JournalArticle"

In the same file I have added the list
DocumentTypeList = ["Novel", "Biography", "Journal Article"]
and
dictionary of "additions_to_reduce_prompt." which is the wording to be added to the reduce_prompt for each document type.
Please do not change this.

So what we could do is (1) determine the document type (2) according to the document type add the appropriate additional prompt to the reduce_prompt.
for document type NOVEL I have this additional reduce prompt:
"Provide a detailed analysis of the novel, including its themes, characters, and plot structure. Insure that you create one structure element for each chapter. In listing characters, locations and organizations list all characters, locations and organizations in the novel but do not list twice if they are reffred to by different names, for example, 'Mr. Darcy' and 'Darcy' should be listed as one character. In each case list the charachter, location or organization by the most common way it is refrred to in the text."

It seems that we need an additional AI call to determine the document type. It would be great if we could make this call to run at at the same time as running the _analyze_large_block elements are running - say by taking the first 10 results that we get of the first 10 calls to _analyze_large_block and using them to make a determination of the document type.

But essentially the idea is to determine the document type and then add text to the reduce_prompt depending on the document type. Can you think of the most efficient way of doing this. I guess one alternative would be to pass in the entire additions_to_reduce_prompt dictionary to the reduce prompt and tell it to using the instructions accoding to the document type it find - but I am not sure how well that would work.