"Cleaned primary entity data works - only done for characters - not organizations or locations. This (1) combines elements which are clearly the same 1.1 Have first and last name the same but one has title and the other not 1.2 Have first and last name and the first names are alternate versions of the same name (e.g. Liz & Betty)
Keeps the name which is a fuller description (e.g Mr Phil Jones, vs Phil Jones - keep "Phil Jones"). If neither is a fuller description keeps the one which is a primary name in more blocks
(2) Removes all primary names where
2.1 The names is used as an alternate name and the names it is used as an alternate name for are in consistant (i.e. could not be referring to the same person)
example: "Miss Smith" is a primary name and is used as an alternate name for the primary names "Jane Smith" and "Mary Smith" - therefore "Miss Smith" refers to more than one
person and so is removed as a primary name
(2) If two of the alternate names are definitely from different people e.g. "Mary Smith" had alternate names "Jane Smith" and "Mary Smith"  are alternate names therefore "Miss Smith" refers to more than one person and so is removed as a primary name.

Next step is to look for "comparison pairs" to determine which primary names refere to the same person